Fixed bugs in ChunkManager and interestListener.

+++ src/ChunkManager.hpp
@@ -5,6 +5,8 @@
 #include <fstream>
 #include <stdio.h>
 #include <string.h>
+#include "TorrentFileManager.hpp"
+
 using namespace std;
 
 class ChunkManager {
@@ -12,25 +14,25 @@ class ChunkManager {
         //Use if bitmap file does not exists
         //If file does not exist, file is preallocated and bitmap is all 0's
         //If file does exists, assumed to be complete and bitmap is all 1's
-        ChunkManager(const char* filename, const long fileSize, const long chunkSize);
+        ChunkManager(const char* filename, const int fileSize, const int chunkSize);
         //Use if bitmap file and file exists
         ChunkManager(const char* filename);
         ~ChunkManager();
-        long readChunk(const long chunkNum, char* data);
-        long writeChunk(const long chunkNum, char* data);
-        long getNumChunks(); //Total number of chunks
-        long getChunkSize(); //Size of chunk, except last
-        long getLastChunkSize(); //Size of last chunk
+        int readChunk(const int chunkNum, char* data);
+        int writeChunk(const int chunkNum, char* data);
+        int getNumChunks(); //Total number of chunks
+        int getChunkSize(); //Size of chunk, except last
+        int getLastChunkSize(); //Size of last chunk
         bool* getChunks(); //Return pointer to bitmap
-        bool chunkAvailable(const long chunkNum); //Yes/No if chunk has been downloaded
+        bool chunkAvailable(const int chunkNum); //Yes/No if chunk has been downloaded
         string getBitstring();
         
     private:
         fstream m_fileStream;
         fstream m_metaFileStream;
-        long m_lastChunkSize;
-        long m_chunkSize;
-        long m_numChunks;
+        int m_lastChunkSize;
+        int m_chunkSize;
+        int m_numChunks;
         bool* m_chunks;
         
         void readMetaFile();
